getting some errors but minimap git diff is working

+++ packages/linter/lib/linter.coffee
@@ -83,6 +83,8 @@ class Linter
     cmd_list = cmd_list.map (cmd_item) ->
       if /@filename/i.test(cmd_item)
         return cmd_item.replace(/@filename/gi, filePath)
+      if /@tempdir/i.test(cmd_item)
+        return cmd_item.replace(/@tempdir/gi, path.dirname(filePath))
       else
         return cmd_item
 
@@ -92,6 +94,9 @@ class Linter
       command: cmd_list[0],
       args: cmd_list.slice(1)
     }
+
+  getReportFilePath: (filePath) ->
+    path.join(path.dirname(filePath), @reportFilePath)
 
   # Private: Provide the node executable path for use when executing a node
   #          linter
@@ -113,6 +118,7 @@ class Linter
 
     dataStdout = []
     dataStderr = []
+    exited = false
 
     stdout = (output) ->
       log 'stdout', output
@@ -123,15 +129,27 @@ class Linter
       dataStderr += output
 
     exit = =>
-      data = if @errorStream is 'stdout' then dataStdout else dataStderr
+      exited = true
+      switch @errorStream
+        when 'file'
+          reportFilePath = @getReportFilePath(filePath)
+          if fs.existsSync reportFilePath
+            data = fs.readFileSync(reportFilePath)
+        when 'stdout' then data = dataStdout
+        else data = dataStderr
       @processMessage data, callback
 
     process = new BufferedProcess({command, args, options,
                                   stdout, stderr, exit})
 
     # Don't block UI more than 5seconds, it's really annoying on big files
+    # TODO: This doesn't actually block a UI thread, but it does cause lint
+    # warnings to flash. A better fix would be to diff new lint messages with
+    # existing ones and remove those that are no longer present. Right now we
+    # just remove all existing ones, and add all new ones.
     timeout_s = 5
     setTimeout ->
+      return if exited
       process.kill()
       warn "command `#{command}` timed out after #{timeout_s}s"
     , timeout_s * 1000
@@ -170,7 +188,15 @@ class Linter
     else
       level = @defaultLevel
 
+    # If no line/col is found, assume a full file error
+    # TODO: This conflicts with the docs above that say line is required :(
+    match.line ?= 0
+    match.col ?= 0
+
     return {
+      # TODO: It's confusing that line & col are here since they duplicate info
+      # that's present in the value for range. Consider deprecating line & col
+      # since they're less general than range.
       line: match.line,
       col: match.col,
       level: level,
@@ -187,7 +213,8 @@ class Linter
     match.message
 
   lineLengthForRow: (row) ->
-    return @editor.lineLengthForBufferRow row
+    text = @editor.lineTextForBufferRow row
+    return text?.length or 0
 
   getEditorScopesForPosition: (position) ->
     try
@@ -222,7 +249,6 @@ class Linter
   #   colStart: column to on which to start a higlight (optional)
   #   colEnd: column to end highlight (optional)
   computeRange: (match) ->
-    match.line ?= 0 # Assume if no line is found that it denotes a full file error.
 
     decrementParse = (x) ->
       Math.max 0, parseInt(x) - 1
@@ -236,7 +262,6 @@ class Linter
       log "ignoring #{match} - it's longer than the buffer"
       return null
 
-    match.col ?=  0
     unless match.colStart
       position = new Point(rowStart, match.col)
       scopes = @getEditorScopesForPosition(position)
